requirements.txt
django~=5.0.1
djangorestframework
drf-spectacular
psycopg2-binary
python-decouple
django-cors-headers
drf-spectacular-sidecar
constructs